ml-simple/requirements-connectors.txt
# Data Connector Dependencies
# Install with: pip install -r requirements-connectors.txt

# Core (required)
pandas>=2.1.4
sqlalchemy>=2.0.0

# Database Drivers
pymysql>=1.1.0              # MySQL
psycopg2-binary>=2.9.0      # PostgreSQL
pyodbc>=5.0.0               # SQL Server
cx_Oracle>=8.3.0            # Oracle
ibm_db>=3.2.0               # IBM DB2

# Cloud Storage
boto3>=1.34.0               # AWS S3
google-cloud-storage>=2.14.0  # Google Cloud Storage
azure-storage-blob>=12.19.0   # Azure Blob Storage

# Data Warehouses
google-cloud-bigquery>=3.14.0           # BigQuery
snowflake-connector-python>=3.6.0       # Snowflake
redshift-connector>=2.0.0               # Redshift (alternative to psycopg2)
databricks-sql-connector>=2.9.0         # Databricks
teradatasql>=17.20.0.0                  # Teradata

# Optional: For better performance
pyarrow>=14.0.0            # Parquet support
fastparquet>=2023.10.0     # Alternative parquet engine

# Optional: For Redshift with psycopg2
# psycopg2-binary>=2.9.0   # Already included above for PostgreSQL
